interrupção

+++ Core/Src/general_can.c
@@ -9,9 +9,8 @@ static FDCAN_TxHeaderTypeDef TxHeader;
 
 uint8_t RxData[8];
 FDCAN_RxHeaderTypeDef RxHeader;
-int16_t datageneral[4];
+uint16_t datageneral[4];
 uint32_t idgeneral;
-
 
 
 //função que inicializa a can geral, chamada em initializer.c
@@ -33,7 +32,6 @@ void general_can_transmit(uint32_t id, uint16_t* data) {
 //função de callback, chamada quando chega qualquer mensagem, de qualquer ID
 void CAN_general_receive_callback(FDCAN_HandleTypeDef* hfdcan, uint32_t RxFifo0ITs)  {
 	if((RxFifo0ITs & FDCAN_IT_RX_FIFO0_NEW_MESSAGE) != RESET) {
-		//piscar LED da CAN PD3
 		if (HAL_FDCAN_GetRxMessage(hfdcan, FDCAN_RX_FIFO0, &RxHeader, RxData) != HAL_OK) {
 			/* Reception Error */
 			Error_Handler();
@@ -41,10 +39,10 @@ void CAN_general_receive_callback(FDCAN_HandleTypeDef* hfdcan, uint32_t RxFifo0I
 
 		idgeneral = RxHeader.Identifier;
 		for(int i = 0; i < 8; i += 2){
-			datageneral[i/2] = (RxData[i+1] << 8) | RxData[i];
+			datageneral[i/2] = (uint16_t)((RxData[i+1] << 8) | RxData[i]);
 		}
 		HAL_GPIO_TogglePin(GPIOD, GPIO_PIN_6); //A LED DA CAN (PIN 3, TÁ QUEIMADA)
-		//TODO: implementar lógica de colocar as mensagens nas variáveis certas
+
 		canMessageReceived(idgeneral, datageneral);
 
 		if (HAL_FDCAN_ActivateNotification(hfdcan, FDCAN_IT_RX_FIFO0_NEW_MESSAGE, 0) != HAL_OK) {
@@ -53,3 +51,5 @@ void CAN_general_receive_callback(FDCAN_HandleTypeDef* hfdcan, uint32_t RxFifo0I
 		}
 	}
 }
+
+
